DTO for send message
